+++ include/swift/Basic/BasicBridging.h
@@ -99,9 +99,47 @@ typedef uintptr_t SwiftUInt;
 #define BRIDGING_WRAPPER_NULLABLE(Name) \
   BRIDGING_WRAPPER_IMPL(Name, Nullable##Name, _Nullable)
 
+//===----------------------------------------------------------------------===//
+// ArrayRef
+//===----------------------------------------------------------------------===//
+
+struct BridgedArrayRef {
+  const void *_Nullable data;
+  size_t numElements;
+};
+
+//===----------------------------------------------------------------------===//
+// Data
+//===----------------------------------------------------------------------===//
+
+typedef struct BridgedData {
+  const char *_Nullable baseAddress;
+  size_t size;
+} BridgedData;
+
+void BridgedData_free(BridgedData data);
+
+//===----------------------------------------------------------------------===//
+// Feature
+//===----------------------------------------------------------------------===//
+
+typedef enum ENUM_EXTENSIBILITY_ATTR(open) BridgedFeature {
+#define LANGUAGE_FEATURE(FeatureName, SENumber, Description, Option)           \
+  FeatureName,
+#include "swift/Basic/Features.def"
+} BridgedFeature;
+
+//===----------------------------------------------------------------------===//
+// OStream
+//===----------------------------------------------------------------------===//
+
 struct BridgedOStream {
   void * _Nonnull streamAddr;
 };
+
+//===----------------------------------------------------------------------===//
+// StringRef
+//===----------------------------------------------------------------------===//
 
 class BridgedStringRef {
   const char *_Nullable Data;
@@ -156,6 +194,10 @@ BRIDGED_INLINE SwiftInt BridgedOwnedString_count(BridgedOwnedString str);
 SWIFT_NAME("getter:BridgedOwnedString.isEmpty(self:)")
 BRIDGED_INLINE bool BridgedOwnedString_empty(BridgedOwnedString str);
 
+//===----------------------------------------------------------------------===//
+// SourceLoc
+//===----------------------------------------------------------------------===//
+
 class BridgedSourceLoc {
   const void *_Nullable Raw;
 
@@ -185,10 +227,85 @@ public:
 SWIFT_NAME("getter:BridgedSourceLoc.isValid(self:)")
 BRIDGED_INLINE bool BridgedSourceLoc_isValid(BridgedSourceLoc str);
 
-struct BridgedArrayRef {
-  const void * _Nullable data;
-  size_t numElements;
-};
+//===----------------------------------------------------------------------===//
+// SourceRange
+//===----------------------------------------------------------------------===//
+
+typedef struct {
+  BridgedSourceLoc startLoc;
+  BridgedSourceLoc endLoc;
+} BridgedSourceRange;
+
+typedef struct {
+  void *_Nonnull start;
+  size_t byteLength;
+} BridgedCharSourceRange;
+
+//===----------------------------------------------------------------------===//
+// Plugins
+//===----------------------------------------------------------------------===//
+
+SWIFT_BEGIN_ASSUME_NONNULL
+
+/// Create a new root 'null' JSON value. Clients must call \c JSON_value_delete
+/// after using it.
+void *JSON_newValue();
+
+/// Parse \p data as a JSON data and return the top-level value. Clients must
+/// call \c JSON_value_delete after using it.
+void *JSON_deserializedValue(BridgedData data);
+
+/// Serialize a value and populate \p result with the result data. Clients
+/// must call \c BridgedData_free after using the \p result.
+void JSON_value_serialize(void *valuePtr, BridgedData *result);
+
+/// Destroy and release the memory for \p valuePtr that is a result from
+/// \c JSON_newValue() or \c JSON_deserializedValue() .
+void JSON_value_delete(void *valuePtr);
+
+bool JSON_value_getAsNull(void *valuePtr);
+bool JSON_value_getAsBoolean(void *valuePtr, bool *result);
+bool JSON_value_getAsString(void *valuePtr, BridgedData *result);
+bool JSON_value_getAsDouble(void *valuePtr, double *result);
+bool JSON_value_getAsInteger(void *valuePtr, int64_t *result);
+bool JSON_value_getAsObject(void *valuePtr, void *_Nullable *_Nonnull result);
+bool JSON_value_getAsArray(void *valuePtr, void *_Nullable *_Nonnull result);
+
+size_t JSON_object_getSize(void *objectPtr);
+BridgedData JSON_object_getKey(void *objectPtr, size_t i);
+bool JSON_object_hasKey(void *objectPtr, const char *key);
+void *JSON_object_getValue(void *objectPtr, const char *key);
+
+size_t JSON_array_getSize(void *arrayPtr);
+void *JSON_array_getValue(void *arrayPtr, size_t index);
+
+void JSON_value_emplaceNull(void *valuePtr);
+void JSON_value_emplaceBoolean(void *valuePtr, bool value);
+void JSON_value_emplaceString(void *valuePtr, const char *value);
+void JSON_value_emplaceDouble(void *valuePtr, double value);
+void JSON_value_emplaceInteger(void *valuePtr, int64_t value);
+void *JSON_value_emplaceNewObject(void *valuePtr);
+void *JSON_value_emplaceNewArray(void *valuePtr);
+
+void JSON_object_setNull(void *objectPtr, const char *key);
+void JSON_object_setBoolean(void *objectPtr, const char *key, bool value);
+void JSON_object_setString(void *objectPtr, const char *key, const char *value);
+void JSON_object_setDouble(void *objectPtr, const char *key, double value);
+void JSON_object_setInteger(void *objectPtr, const char *key, int64_t value);
+void *JSON_object_setNewObject(void *objectPtr, const char *key);
+void *JSON_object_setNewArray(void *objectPtr, const char *key);
+void *JSON_object_setNewValue(void *objectPtr, const char *key);
+
+void JSON_array_pushNull(void *arrayPtr);
+void JSON_array_pushBoolean(void *arrayPtr, bool value);
+void JSON_array_pushString(void *arrayPtr, const char *value);
+void JSON_array_pushDouble(void *arrayPtr, double value);
+void JSON_array_pushInteger(void *arrayPtr, int64_t value);
+void *JSON_array_pushNewObject(void *arrayPtr);
+void *JSON_array_pushNewArray(void *arrayPtr);
+void *JSON_array_pushNewValue(void *arrayPtr);
+
+SWIFT_END_ASSUME_NONNULL
 
 SWIFT_END_NULLABILITY_ANNOTATIONS
