+++ lib/ASTGen/Sources/ASTGen/DiagnosticsBridge.swift
@@ -1,6 +1,5 @@
 import ASTBridging
 import BasicBridging
-import CASTBridging
 import SwiftDiagnostics
 import SwiftSyntax
