+++ lib/ASTGen/Sources/ASTGen/ParameterClause.swift
@@ -10,8 +10,8 @@
 //
 //===----------------------------------------------------------------------===//
 
+import ASTBridging
 import BasicBridging
-import CASTBridging
 import SwiftSyntax
 
 // MARK: - ParamDecl
