+++ lib/ASTGen/Sources/ASTGen/SourceManager.swift
@@ -1,6 +1,5 @@
 import ASTBridging
 import BasicBridging
-import CASTBridging
 import SwiftOperators
 import SwiftSyntax
 import SwiftSyntaxMacros
