+++ lib/Basic/CMakeLists.txt
@@ -45,9 +45,7 @@ add_swift_host_library(swiftBasic STATIC
   BasicBridging.cpp
   BasicSourceInfo.cpp
   Cache.cpp
-  CBasicBridging.cpp
   ClusteredBitVector.cpp
-  CBasicBridging.cpp
   DiverseStack.cpp
   Edit.cpp
   EditorPlaceholder.cpp
